requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pkg-resources==0.0.0
python-dotenv==0.12.0
requests==2.23.0
urllib3==1.25.8
